Keeping Whole Message
Previously, when we kept two columns, it disregarded all the information after the second comma, getting rid of some of our data. In this commit, I set it up to print the first column, then delete the first column and print the rest of the row - so we don't lose any data.